Move tasktree into expandable element
